feat: Add help and remove some warnings

+++ src/2d/hud.cpp
@@ -41,8 +41,8 @@ void Hud::update() {
 
 			if (glo::game->_cam_index >= 1 && glo::game->_cam_index <= 4) {
 				Texture* tex_cam_hud = glo::res->load_tex("tex_cam_hud.png");
-				float cam_toggle_x = (w - tex_cam_hud->w) / 2;
-				float cam_toggle_y = (h - tex_cam_hud->h - 32);
+				float cam_toggle_x = (w - tex_cam_hud->w) / 2.0f;
+				float cam_toggle_y = (h - tex_cam_hud->h - 32.0f);
 
 				BBox bbox = BBox({ cam_toggle_x , cam_toggle_y }, { tex_cam_hud->w, tex_cam_hud->h });
 				if (bbox.contains(input.mp_curr)) {
@@ -72,13 +72,13 @@ void Hud::draw(Shader* shd) {
 		shd->set_mat4("u_proj", &proj[0][0]);
 		shd->set_vec2("u_pos", hud_x, hud_y);
 		shd->set_vec2("u_scale", map.tex->w * scale, map.tex->h * scale);
-		shd->set_vec3("u_img_frame", map.tex->w / (float)map.tex->fullw, map.tex->h / (float)map.tex->fullh, frame);
+		shd->set_vec3("u_img_frame", map.tex->w / (float)map.tex->fullw, map.tex->h / (float)map.tex->fullh, (float)frame);
 		shd->set_vec4("u_color", 1, 1, 1, 1);
 		mesh_2d->draw(shd, map.tex);
 
 		Texture* tex_cam_hud = glo::res->load_tex("tex_cam_hud.png");
-		float cam_toggle_x = (w - tex_cam_hud->w) / 2;
-		float cam_toggle_y = (h - tex_cam_hud->h - 32);
+		float cam_toggle_x = (w - tex_cam_hud->w) / 2.0f;
+		float cam_toggle_y = (h - tex_cam_hud->h - 32.0f);
 		shd->set_mat4("u_proj", &proj[0][0]);
 		shd->set_vec2("u_pos", cam_toggle_x, cam_toggle_y);
 		shd->set_vec2("u_scale", tex_cam_hud->w * scale, tex_cam_hud->h * scale);
